Move classes to separate .h/.cpp files
