+++ esphome/components/select/select.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include <set>
-#include <utility>
 #include "esphome/core/component.h"
 #include "esphome/core/entity_base.h"
 #include "esphome/core/helpers.h"
+#include "select_call.h"
+#include "select_traits.h"
 
 namespace esphome {
 namespace select {
@@ -16,67 +16,6 @@ namespace select {
       ESP_LOGCONFIG(TAG, "%s  Icon: '%s'", prefix, (obj)->get_icon().c_str()); \
     } \
   }
-
-class Select;
-
-enum SelectOp {
-  SELECT_OP_NONE,
-  SELECT_OP_SET,
-  SELECT_OP_NEXT,
-  SELECT_OP_PREVIOUS,
-  SELECT_OP_FIRST,
-  SELECT_OP_LAST
-};
-
-class SelectCall {
- public:
-  explicit SelectCall(Select *parent) : parent_(parent) {}
-  void perform();
-
-  SelectCall &set_option(const std::string &option) {
-    op_ = SELECT_OP_SET;
-    option_ = option;
-    return *this;
-  }
-  const optional<std::string> &get_option() const { return option_; }
-
-  SelectCall &select_next(bool cycle) {
-    op_ = SELECT_OP_NEXT;
-    cycle_ = cycle;
-    return *this;
-  }
-
-  SelectCall &select_previous(bool cycle) {
-    op_ = SELECT_OP_PREVIOUS;
-    cycle_ = cycle;
-    return *this;
-  }
-
-  SelectCall &select_first() {
-    op_ = SELECT_OP_FIRST;
-    return *this;
-  }
-
-  SelectCall &select_last() {
-    op_ = SELECT_OP_LAST;
-    return *this;
-  }
-
- protected:
-  Select *const parent_;
-  SelectOp op_{SELECT_OP_NONE};
-  optional<std::string> option_;
-  bool cycle_;
-};
-
-class SelectTraits {
- public:
-  void set_options(std::vector<std::string> options) { this->options_ = std::move(options); }
-  std::vector<std::string> get_options() const { return this->options_; }
-
- protected:
-  std::vector<std::string> options_;
-};
 
 /** Base-class for all selects.
  *
