+++ esphome/components/select/select_call.h
@@ -1,0 +1,40 @@
+#pragma once
+
+#include "esphome/core/helpers.h"
+
+namespace esphome {
+namespace select {
+
+class Select;
+
+enum SelectOp {
+  SELECT_OP_NONE,
+  SELECT_OP_SET,
+  SELECT_OP_NEXT,
+  SELECT_OP_PREVIOUS,
+  SELECT_OP_FIRST,
+  SELECT_OP_LAST
+};
+
+class SelectCall {
+ public:
+  explicit SelectCall(Select *parent) : parent_(parent) {}
+  void perform();
+
+  SelectCall &set_option(const std::string &option);
+  const optional<std::string> &get_option() const;
+
+  SelectCall &select_next(bool cycle);
+  SelectCall &select_previous(bool cycle);
+  SelectCall &select_first();
+  SelectCall &select_last();
+
+ protected:
+  Select *const parent_;
+  SelectOp op_{SELECT_OP_NONE};
+  optional<std::string> option_;
+  bool cycle_;
+};
+
+}  // namespace select
+}  // namespace esphome
